1) Moving models to a separate file for Customer and Customer User 2) Typo fix 3) Changed customerUser file name to customer_user
